+++ Sources/NexusFramework/Time/Time.h
@@ -30,7 +30,7 @@ namespace NxFr
 
 		NEXUS_FRAMEWORK_API Timestamp Now();
 		NEXUS_FRAMEWORK_API int64 TimeSinceEpoch();
-		NEXUS_FRAMEWORK_API uint64 ProcessorTick();
+		NEXUS_FRAMEWORK_API double ProcessorTick();
 		NEXUS_FRAMEWORK_API double Convert(double Value, double Unit = 1.0);
 
 		NEXUS_FRAMEWORK_API StringView GetWeekDay(int32 Day);
